% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybryd-subset-selection.R
\name{computePCEscore}
\alias{computePCEscore}
\title{computePCEscore}
\usage{
computePCEscore(data)
}
\arguments{
\item{data}{a dataframe with 3 columns: x-axis coordinates (labeled `x`), y-axis coordinates(labeled `y`), and the class labels (labeled as `labels`)}
}
\description{
computePCEscore: computes the pixel class entropy score for any biaxial dataset with class labels
}
