% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid-subset-selection.R
\name{create_pixel_grid}
\alias{create_pixel_grid}
\title{create_pixel_grid}
\usage{
create_pixel_grid(xbreaks = 100, ybreaks = 100)
}
\arguments{
\item{xbreaks}{the # of pixels to break the x-axis into. Defaults to 100.}

\item{ybreaks}{the # of pixels to break the y-axis into. Defaults to 100.}
}
\description{
create_pixel_grid: This function generates a pixel grid template, defaults to 10,000 pixels
}
\keyword{internal}
