% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridSubsetSelection.R
\name{makeAxes}
\alias{makeAxes}
\title{makeAxes}
\usage{
makeAxes(df, co = coefficients)
}
\arguments{
\item{df}{a dataframe of cells (rows) by markers/genes (columns)}

\item{co}{the dataframe of LDA coefficients from MASS::lda.}
}
\description{
makeAxes: function that generates new HSS-LDA axes given an LDA model coefficients
}
\keyword{internal}
