% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TcellHartmann2020_sampleData}
\alias{TcellHartmann2020_sampleData}
\title{T-cell metabolic regulome dataset}
\format{
A data frame of 6000 rows and 16 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/32868913/}
}
\usage{
data(TcellHartmann2020_sampleData)
}
\description{
Single-cell data from an in vitro TCR stimulation experiment where cells are extracted at different days from 0-5 and analyzed using CyTOF.
1000 cells per timepoint, and only 15 metabolic markers
}
\references{
Hartmann et al. (2021) Nat Biotechnol 39, 186–197
}
\keyword{datasets}
